# apps/tsk-api/openapi.json
{"openapi":"3.0.0","paths":{"/v1":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/v1/auth/sign-up/email":{"post":{"operationId":"AuthController_setEmailAndPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterSetEmailAndPassword"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetEmailAndPasswordHttpResponse"}}}},"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpDefaultResponse"}}}}},"tags":["auth"],"security":[{"uid":[]}]}},"/v1/auth/verify/email":{"post":{"operationId":"AuthController_verifyEmail","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyPhoneNumberOrEmailAddressDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OtpVerificationHttpResponse"}}}},"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpDefaultResponse"}}}}},"tags":["auth"],"security":[{"uid":[]},{"bearer":[]}]}},"/v1/auth/login/request":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginHttpResponse"}}}},"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpDefaultResponse"}}}}},"tags":["auth"],"security":[{"uid":[]}]}},"/v1/auth/login/2fa-verification":{"post":{"operationId":"AuthController_verifyLogin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyPhoneNumberOrEmailAddressDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OtpVerificationHttpResponse"}}}},"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpDefaultResponse"}}}}},"tags":["auth"],"security":[{"uid":[]},{"bearer":[]}]}},"/v1/auth/logout/kill-session":{"delete":{"operationId":"AuthController_logout","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LogoutHttpReponse"}}}},"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpDefaultResponse"}}}}},"tags":["auth"],"security":[{"uid":[]},{"bearer":[]}]}},"/v1/project/create":{"post":{"operationId":"ProjectController_createProject","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProjectRequestPayload"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectResponse"}}}},"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpDefaultResponse"}}}}},"tags":["project"],"security":[{"uid":[]},{"bearer":[]}]}},"/v1/project":{"get":{"operationId":"ProjectController_getProject","parameters":[{"name":"projectId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectResponse"}}}},"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpDefaultResponse"}}}}},"tags":["project"],"security":[{"uid":[]},{"bearer":[]}]}},"/v1/project/projects":{"get":{"operationId":"ProjectController_getProjects","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectsResponse"}}}},"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpDefaultResponse"}}}}},"tags":["project"],"security":[{"uid":[]},{"bearer":[]}]}},"/v1/task/create":{"post":{"operationId":"TaskController_addTask","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTaskRequestPayload"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskResponse"}}}},"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpDefaultResponse"}}}}},"tags":["task"],"security":[{"uid":[]},{"bearer":[]}]}},"/v1/task/state":{"put":{"operationId":"TaskController_updateTaskState","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeTaskRequestPayload"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpBooleanResponse"}}}},"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpDefaultResponse"}}}}},"tags":["task"],"security":[{"uid":[]},{"bearer":[]}]}},"/v1/task":{"delete":{"operationId":"TaskController_deleteTask","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteTaskRequestPayload"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpBooleanResponse"}}}},"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpDefaultResponse"}}}}},"tags":["task"],"security":[{"uid":[]},{"bearer":[]}]}}},"info":{"title":"@tsk-man/api API","description":"","version":"0.0.1","contact":{}},"tags":[{"name":"x-service-port","description":"3080"}],"servers":[],"components":{"securitySchemes":{"uid":{"type":"apiKey","name":"x-uid","in":"header"},"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"UserDeviceDto":{"type":"object","properties":{"type":{"type":"string","example":"mobile","description":"Type of the device (e.g., mobile, desktop)"},"ip":{"type":"string","example":"user Ip","description":"user Ip address"},"model":{"type":"string","example":"iPhone 12","description":"Model of the device"},"os":{"type":"string","example":"iOS","description":"Operating system of the device"},"osVersion":{"type":"string","example":"12","description":"Operating system version of the device"},"serialNumber":{"type":"string","example":"123456789","description":"Serial number of the device"},"manufacturer":{"type":"string","example":"Apple","description":"Manufacturer of the device"},"isPhysicalDevice":{"type":"boolean","example":false},"packageName":{"type":"string","example":"com.tsk-man.user"},"appVersion":{"type":"string","example":"v1.0.2"}}},"RegisterSetEmailAndPassword":{"type":"object","properties":{"emailAddr":{"type":"string","example":"test@gmail.com","description":"The email of the User","format":"email","uniqueItems":true,"minLength":5,"maxLength":255},"password":{"type":"string","example":"secret password change me!","description":"The password of the User","format":"string","minLength":6,"maxLength":32},"confirmPassword":{"type":"string","example":"secret password change me!","description":"The password of the User","format":"string","minLength":6,"maxLength":32},"firstName":{"type":"string","example":"Kamsi","format":"string","minLength":2},"lastName":{"type":"string","example":"Mofe","format":"string","minLength":2},"deviceDetails":{"description":"An Object mapping user device details to it corresponding details","allOf":[{"$ref":"#/components/schemas/UserDeviceDto"}]}},"required":["emailAddr","password","confirmPassword","firstName","lastName","deviceDetails"]},"SetEmailAndPasswordResponse":{"type":"object","properties":{"token":{"type":"string"},"userId":{"type":"string"},"next":{"type":"string","enum":["email_verification","t2fa","dashboard"]}},"required":["token","userId","next"]},"SetEmailAndPasswordHttpResponse":{"type":"object","properties":{"isError":{"type":"boolean","description":"Indicates if the request was successful or an error"},"description":{"type":"string","description":"A details error message for developers. Will be an empty string in production"},"context":{"type":"string"},"message":{"type":"boolean","description":"A friendly message for the user. In a error it gives the user information about the error and how to resolve the errors"},"payload":{"nullable":true,"description":"A distionary that maps the body of the response to it coressponding values","allOf":[{"$ref":"#/components/schemas/SetEmailAndPasswordResponse"}]}},"required":["isError","description","context","message","payload"]},"HttpDefaultResponse":{"type":"object","properties":{"isError":{"type":"boolean","description":"Indicates if the request was successful or an error"},"description":{"type":"string","description":"A details error message for developers. Will be an empty string in production"},"context":{"type":"string"},"message":{"type":"boolean","description":"A friendly message for the user. In a error it gives the user information about the error and how to resolve the errors"},"payload":{"type":"object","nullable":true,"description":"A distionary that maps the body of the response to it coressponding values"}},"required":["isError","description","context","message","payload"]},"VerifyPhoneNumberOrEmailAddressDto":{"type":"object","properties":{"otp":{"type":"string","example":"otp","description":"A six digit character that was sent to the identity been verified"},"identity":{"type":"string","example":"+23408020144002 || example@gmail.com","description":"The phone number or email address been verified"},"userId":{"type":"string","description":"A string representation for identifing user"},"deviceDetails":{"description":"An Object mapping user device details to it corresponding details","allOf":[{"$ref":"#/components/schemas/UserDeviceDto"}]}},"required":["otp","identity","userId","deviceDetails"]},"OtpVerificationResponse":{"type":"object","properties":{"token":{"type":"string","description":"Json Web Token for validating next step validation "},"userId":{"type":"string","description":"User Identifier"},"next":{"type":"string","enum":["email_verification","t2fa","dashboard"],"description":"The next step/screen that should be shown to the user on the client"}},"required":["token","userId","next"]},"OtpVerificationHttpResponse":{"type":"object","properties":{"isError":{"type":"boolean","description":"Indicates if the request was successful or an error"},"description":{"type":"string","description":"A details error message for developers. Will be an empty string in production"},"context":{"type":"string"},"message":{"type":"boolean","description":"A friendly message for the user. In a error it gives the user information about the error and how to resolve the errors"},"payload":{"nullable":true,"description":"A distionary that maps the body of the response to it coressponding values","allOf":[{"$ref":"#/components/schemas/OtpVerificationResponse"}]}},"required":["isError","description","context","message","payload"]},"LoginRequest":{"type":"object","properties":{"email":{"type":"string","example":"example@gmail.com"},"password":{"type":"string","example":"qaws3ed!"},"device":{"description":"Details about the device this user is making request from","allOf":[{"$ref":"#/components/schemas/UserDeviceDto"}]}},"required":["email","password","device"]},"LoginResponse":{"type":"object","properties":{"token":{"type":"string"},"userId":{"type":"string"},"next":{"type":"string","enum":["email_verification","t2fa","dashboard"]}},"required":["token","userId","next"]},"LoginHttpResponse":{"type":"object","properties":{"isError":{"type":"boolean","description":"Indicates if the request was successful or an error"},"description":{"type":"string","description":"A details error message for developers. Will be an empty string in production"},"context":{"type":"string"},"message":{"type":"boolean","description":"A friendly message for the user. In a error it gives the user information about the error and how to resolve the errors"},"payload":{"nullable":true,"description":"A distionary that maps the body of the response to it coressponding values","allOf":[{"$ref":"#/components/schemas/LoginResponse"}]}},"required":["isError","description","context","message","payload"]},"LogoutHttpReponse":{"type":"object","properties":{"isError":{"type":"boolean","description":"Indicates if the request was successful or an error"},"description":{"type":"string","description":"A details error message for developers. Will be an empty string in production"},"context":{"type":"string"},"message":{"type":"boolean","description":"A friendly message for the user. In a error it gives the user information about the error and how to resolve the errors"},"payload":{"type":"object","nullable":true,"description":"boolean value representing if the session was successfully killed"}},"required":["isError","description","context","message","payload"]},"CreateProjectRequestPayload":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"}},"required":["name","description"]},"ProjectDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"status":{"type":"string","enum":["deleted","active"]},"projectId":{"type":"string","description":"Unique identifier for the project"},"createdBy":{"type":"string","description":"Unique identifier of the user who created this project"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["name","description","status","projectId","createdBy","createdAt","updatedAt"]},"ProjectResponse":{"type":"object","properties":{"isError":{"type":"boolean","description":"Indicates if the request was successful or an error"},"description":{"type":"string","description":"A details error message for developers. Will be an empty string in production"},"context":{"type":"string"},"message":{"type":"boolean","description":"A friendly message for the user. In a error it gives the user information about the error and how to resolve the errors"},"payload":{"nullable":true,"description":"A map matching project details to it corresponding values","allOf":[{"$ref":"#/components/schemas/ProjectDto"}]}},"required":["isError","description","context","message","payload"]},"ProjectsWithPagination":{"type":"object","properties":{"projects":{"description":"An array of projects with it corresponding values","type":"array","items":{"$ref":"#/components/schemas/ProjectDto"}},"paginationCursor":{"type":"object","nullable":true,"description":"An object that stores the the position of the client while navigating through a large project list (TODO)"}},"required":["projects","paginationCursor"]},"ProjectsResponse":{"type":"object","properties":{"isError":{"type":"boolean","description":"Indicates if the request was successful or an error"},"description":{"type":"string","description":"A details error message for developers. Will be an empty string in production"},"context":{"type":"string"},"message":{"type":"boolean","description":"A friendly message for the user. In a error it gives the user information about the error and how to resolve the errors"},"payload":{"nullable":true,"description":"","allOf":[{"$ref":"#/components/schemas/ProjectsWithPagination"}]}},"required":["isError","description","context","message","payload"]},"CreateTaskRequestPayload":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"dueDate":{"type":"string"},"projectId":{"type":"string"},"tags":{"type":"array","items":{"type":"array"}}},"required":["title","description","dueDate","projectId","tags"]},"TaskDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"dueDate":{"type":"string"},"projectId":{"type":"string"},"tags":{"type":"array","items":{"type":"array"}},"taskId":{"type":"string"},"authorId":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"state":{"type":"string","enum":["todo","progress","done"]}},"required":["title","description","dueDate","projectId","tags","taskId","authorId","createdAt","updatedAt","state"]},"TaskResponse":{"type":"object","properties":{"isError":{"type":"boolean","description":"Indicates if the request was successful or an error"},"description":{"type":"string","description":"A details error message for developers. Will be an empty string in production"},"context":{"type":"string"},"message":{"type":"boolean","description":"A friendly message for the user. In a error it gives the user information about the error and how to resolve the errors"},"payload":{"nullable":true,"description":"A distionary that maps the body of the response to it coressponding values","allOf":[{"$ref":"#/components/schemas/TaskDto"}]}},"required":["isError","description","context","message","payload"]},"ChangeTaskRequestPayload":{"type":"object","properties":{"projectId":{"type":"string"},"taskId":{"type":"string"},"taskState":{"type":"string","enum":["todo","progress","done"]}},"required":["projectId","taskId","taskState"]},"HttpBooleanResponse":{"type":"object","properties":{"isError":{"type":"boolean","description":"Indicates if the request was successful or an error"},"description":{"type":"string","description":"A details error message for developers. Will be an empty string in production"},"context":{"type":"string"},"message":{"type":"boolean","description":"A friendly message for the user. In a error it gives the user information about the error and how to resolve the errors"},"payload":{"type":"boolean","nullable":true,"description":"A distionary that maps the body of the response to it coressponding values"}},"required":["isError","description","context","message","payload"]},"DeleteTaskRequestPayload":{"type":"object","properties":{"projectId":{"type":"string"},"taskId":{"type":"string"}},"required":["projectId","taskId"]}}}}